Fikset feil i scheduled tasks, flyttet på noen ting, enablet kafa listener og scheduled tasks i test-context
