+++ src/main/kotlin/no/nav/arbeidsplassen/importapi/config/ApplicationProperties.kt
@@ -30,10 +30,6 @@ class NaisSecretSignatureConfigProperties(env: Map<String, String>) : SecretSign
     secret = env.variable("JWT_SECRET")
 )
 
-class TestSecretSignatureConfigProperties() : SecretSignatureConfigProperties(
-    secret = "Thisisaverylongsecretandcanonlybeusedintest",
-)
-
 abstract class DatabaseConfigProperties(
     val host: String,
     val port: String,
@@ -56,21 +52,6 @@ class NaisDatabaseConfigProperties(env: Map<String, String>) : DatabaseConfigPro
     additionalParameter = env["DB_ADDITIONAL_PARAMETER"] ?: ""
 )
 
-class TestDatabaseConfigProperties(
-    host: String, port: String
-) : DatabaseConfigProperties(
-    host = host,
-    port = port,
-    database = "test",
-    user = "test",
-    pw = "test",
-    dbDriver = "org.postgresql.Driver",
-    // url = env.variable("DB_URL"),
-    additionalParameter = ""
-) {
-    constructor(env: Map<String, String>) :
-            this(host = env.variable("DB_HOST"), port = env.variable("DB_PORT"))
-}
 
 abstract class KafkaConfigProperties(
     val brokers: String,
@@ -88,18 +69,6 @@ class NaisKafkaConfigProperties(env: Map<String, String>) : KafkaConfigPropertie
     applicationName = env.nullableVariable("NAIS_APP_NAME")
 )
 
-class TestKafkaConfigProperties(
-    brokers: String
-) : KafkaConfigProperties(
-    brokers = brokers,
-    credstorePassword = null,
-    truststorePath = null,
-    keystorePath = null,
-    applicationName = null
-) {
-    constructor(env: Map<String, String>) : this(env.variable("KAFKA_BROKERS"))
-}
-
 abstract class OutgoingPortsConfigProperties(
     val ontologiBaseUrl: String,
     val electorPath: String,
@@ -110,14 +79,6 @@ class NaisOutgoingPortsConfigProperties(env: Map<String, String>) : OutgoingPort
     electorPath = env.variable("ELECTOR_PATH"),
 )
 
-class TestOutgoingPortsConfigProperties(
-    ontologiBaseUrl: String = "http://pam-ontologi", // TODO Denne vil vel aldri være i bruk ?
-    electorPath: String = "NOLEADERELECTION",
-) : OutgoingPortsConfigProperties(
-    ontologiBaseUrl = ontologiBaseUrl,
-    electorPath = electorPath,
-)
-
 abstract class ControllerConfigProperties(
     val adPreviewUrl: String,
     val transferlogBatchSize: Int = 100,
@@ -126,14 +87,6 @@ abstract class ControllerConfigProperties(
 class NaisControllerConfigProperties(env: Map<String, String>) : ControllerConfigProperties(
     adPreviewUrl = env.variable("AD_PREVIEW_URL"),
     transferlogBatchSize = 100
-)
-
-class TestControllerConfigProperties(
-    adPreviewUrl: String = "http://localhost:8080/stillinger/intern",
-    transferlogBatchSize: Int = 100,
-) : ControllerConfigProperties(
-    adPreviewUrl = adPreviewUrl,
-    transferlogBatchSize = transferlogBatchSize
 )
 
 abstract class ServicesConfigProperties(
@@ -154,20 +107,6 @@ class NaisServicesConfigProperties(env: Map<String, String>) : ServicesConfigPro
     adOutboxKafkaProducerTopic = env.variable("ANNONSEMOTTAK_TOPIC"),
 )
 
-class TestServicesConfigProperties(
-    adminStatusPreviewUrl: String = "http://localhost:8080/stillinger/intern",
-    adminStatusTopic: String = "teampam.stilling-intern-1",
-    adminStatusGroupId: String = "import-api-adminstatussync-gcp",
-    adminStatusSyncKafkaEnabled: Boolean = false,
-    adOutboxKafkaProducerTopic: String = "teampam.annonsemottak-1"
-) : ServicesConfigProperties(
-    adminStatusPreviewUrl = adminStatusPreviewUrl,
-    adminStatusTopic = adminStatusTopic,
-    adminStatusGroupId = adminStatusGroupId,
-    adminStatusSyncKafkaEnabled = adminStatusSyncKafkaEnabled,
-    adOutboxKafkaProducerTopic = adOutboxKafkaProducerTopic,
-)
-
 abstract class SchedulerConfigProperties(
     val adOutboxJobEnabled: Boolean,
     val transferlogJobEnabled: Boolean,
@@ -177,12 +116,4 @@ class NaisSchedulerConfigProperties(env: Map<String, String>) : SchedulerConfigP
     adOutboxJobEnabled = env.variable("AD_OUTBOX_SCHEDULER_ENABLED").toBoolean(),
     // Denne mangler i oppsettet, så her vil defaulten bli brukt:
     transferlogJobEnabled = env.nullableVariable("AD_OUTBOX_SCHEDULER_ENABLED")?.toBoolean() ?: true,
-)
-
-class TestSchedulerConfigProperties(
-    adOutboxJobEnabled: Boolean = false,
-    transferlogJobEnabled: Boolean = false
-) : SchedulerConfigProperties(
-    adOutboxJobEnabled = adOutboxJobEnabled,
-    transferlogJobEnabled = transferlogJobEnabled,
 )
