+++ src/main/kotlin/no/nav/arbeidsplassen/importapi/config/SchedulerApplicationContext.kt
@@ -1,7 +1,10 @@
 package no.nav.arbeidsplassen.importapi.config
 
 import no.nav.arbeidsplassen.importapi.scheduler.AdOutboxJob
+import no.nav.arbeidsplassen.importapi.scheduler.AdPulsJob
+import no.nav.arbeidsplassen.importapi.scheduler.DeleteTransferLogJob
 import no.nav.arbeidsplassen.importapi.scheduler.JavalinJobFactory
+import no.nav.arbeidsplassen.importapi.scheduler.TransferLogJob
 import org.quartz.CronScheduleBuilder
 import org.quartz.JobBuilder
 import org.quartz.Scheduler
@@ -12,7 +15,7 @@ import org.slf4j.Logger
 import org.slf4j.LoggerFactory
 
 class SchedulerApplicationContext(
-    schedulerConfigProperties: SchedulerConfigProperties,
+    val schedulerConfigProperties: SchedulerConfigProperties,
     servicesApplicationContext: ServicesApplicationContext
 ) : OnServerStartup, OnServerShutdown {
     companion object {
@@ -30,7 +33,7 @@ class SchedulerApplicationContext(
         .build()
 
 
-    private val adPulsJob = JobBuilder.newJob().ofType(AdOutboxJob::class.java)
+    private val adPulsJob = JobBuilder.newJob().ofType(AdPulsJob::class.java)
         .withIdentity("adPulsJob", "adPuls")
         .build()
 
@@ -40,7 +43,7 @@ class SchedulerApplicationContext(
         .withSchedule(CronScheduleBuilder.cronSchedule("5 15 1 * * ?"))
         .build()
 
-    private val transferLogJob = JobBuilder.newJob().ofType(AdOutboxJob::class.java)
+    private val transferLogJob = JobBuilder.newJob().ofType(TransferLogJob::class.java)
         .withIdentity("transferLogJob", "transferLog")
         .build()
 
@@ -49,7 +52,7 @@ class SchedulerApplicationContext(
         .withSchedule(SimpleScheduleBuilder.repeatSecondlyForever(30))
         .build()
 
-    private val deleteTransferLogJob = JobBuilder.newJob().ofType(AdOutboxJob::class.java)
+    private val deleteTransferLogJob = JobBuilder.newJob().ofType(DeleteTransferLogJob::class.java)
         .withIdentity("deleteTransferLogJob", "deleteTransferLog")
         .build()
 
@@ -70,20 +73,28 @@ class SchedulerApplicationContext(
         LOG.info("Creating scheduler")
         scheduler.setJobFactory(jobFactory)
         if (schedulerConfigProperties.adOutboxJobEnabled) {
+            LOG.info("Scheduling Ad Outbox Job")
             scheduler.scheduleJob(adOutboxJob, adOutboxTrigger)
+        } else {
+            LOG.info("NOT scheduling Ad Outbox Job")
         }
         scheduler.scheduleJob(adPulsJob, adPulsTrigger)
         if (schedulerConfigProperties.transferlogJobEnabled) {
+            LOG.info("Scheduling Transferlog Jobs")
             scheduler.scheduleJob(transferLogJob, transferLogTrigger)
             scheduler.scheduleJob(deleteTransferLogJob, deleteTransferLogTrigger)
+        } else {
+            LOG.info("NOT scheduling Transferlog Jobs")
         }
     }
 
     override fun onServerStartup() {
         LOG.info("Starter scheduler")
         scheduler.start()
-        LOG.info("Trigger jobb ved oppstart")
-        // scheduler.triggerJob(adOutboxJob.key)
+        if (schedulerConfigProperties.adOutboxJobEnabled) {
+            LOG.info("Trigger jobb ved oppstart")
+            scheduler.triggerJob(adOutboxJob.key)
+        }
     }
 
     override fun onServerShutdown() {
