+++ src/test/kotlin/no/nav/arbeidsplassen/importapi/adpuls/AdPulsRepositoryTest.kt
@@ -1,7 +1,7 @@
 package no.nav.arbeidsplassen.importapi.adpuls
 
 import java.util.UUID
-import no.nav.arbeidsplassen.importapi.app.test.TestRepositories
+import no.nav.arbeidsplassen.importapi.app.TestRepositories
 import no.nav.arbeidsplassen.importapi.dao.newTestProvider
 import no.nav.arbeidsplassen.importapi.provider.ProviderRepository
 import no.nav.arbeidsplassen.importapi.repository.TxTemplate
