+++ src/test/kotlin/no/nav/arbeidsplassen/importapi/adstate/AdStateRepositoryTest.kt
@@ -1,7 +1,7 @@
 package no.nav.arbeidsplassen.importapi.adstate
 
 import com.fasterxml.jackson.databind.ObjectMapper
-import no.nav.arbeidsplassen.importapi.app.test.TestRepositories
+import no.nav.arbeidsplassen.importapi.app.TestRepositories
 import no.nav.arbeidsplassen.importapi.dao.newTestProvider
 import no.nav.arbeidsplassen.importapi.dao.transferToAdList
 import no.nav.arbeidsplassen.importapi.provider.ProviderRepository
